// src/styles/theme/type.d.ts
export type ThemeType = {
  colors: {
    brightBlue: string;
    background: {
      linearGradient: string;
    };
    neutrals: {
      primary: string;
      secundary: string;
      terdiary: string;
      light: string;
      semi: string;
      thin: string;
    };
    text: string;
  };
  fonts: {
    josefin: string;
  };
};